[adjustment][case] made a separate component folder for the AddModal feature
